refactor(BookCatalog): move BookInfo to new module
